% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBICgolazo.R
\name{EBICgolazo}
\alias{EBICgolazo}
\title{Finds  the optimal penalty parameter for the GOLAZO approach with respect to the EBIC criterion.}
\usage{
EBICgolazo(
  S,
  n = NULL,
  L = NULL,
  U = NULL,
  tol = 1e-06,
  edge.tol = 1e-06,
  gamma = 0.5,
  rhomin = 0.01,
  rhomax = 1,
  nrhos = 10,
  verbose = TRUE
)
}
\arguments{
\item{S}{a positive semidefinite  matrix}

\item{n}{the sample size}

\item{L}{matrix of lower penalties (may contain -Inf)}

\item{U}{matrix of upper penalties (may contain Inf)}

\item{tol}{the convergence tolerance (default tol=1e-8) for the dual gap}

\item{edge.tol}{the threshold for the normalized entries of K to be treated treated as zero}

\item{gamma}{the EBIC parameter between 0 and 1 (default gamma=0.5)}

\item{rhomin}{minimal rho to check}

\item{rhomax}{maximal rho to check}

\item{nrhos}{number of rhos in the interval (rhomin,rhomax)}

\item{verbose}{if TRUE (default) the output will be printed.}
}
\value{
the  list of rhos for which EBIC was computed

the list of teh corresponding EBIC values

the optimal rho
}
\description{
This function computes the EBIC criterion for a seried of rho parameters. The penalty matrices are \eqn{\rho L}{rho L}, \eqn{\rho U}{rho U},
where L and U are fixed in  advance. If L, U are not specified, they will be set to the default value for the positive GOLAZO problem,
that is, \eqn{L_{ij}=0}{L_ij=0} and \eqn{U_{ij}=1}{U_ij=1} for all \eqn{i\neq j}{i!=j}.
}
\examples{
EBICgolazo(diag(10),n=10)

}
\keyword{EBIC,}
\keyword{optimal}
\keyword{penalty}
