% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golazo.R
\name{golazo}
\alias{golazo}
\title{Performs Graphical Oriented LAZy Optimization by optimizing the dual problem.}
\usage{
golazo(S, L, U, tol = 1e-07, verbose = TRUE)
}
\arguments{
\item{S}{the sample covariance matrix}

\item{L}{matrix of lower penalties (can be -Inf)}

\item{U}{matrix of upper penalties (can be Inf)}

\item{tol}{the convergence tolerance (default tol=1e-7)}

\item{verbose}{if TRUE (default) the output will be printed.}
}
\value{
the optimal value of the concentration matrix

the number of iterations the algorithm needed to converge

the corresponding value of the log-likelihood
}
\description{
This function implements a simple block-coordinate descent algorithm to find the maximum of the regularized
Gaussiann log-likelihood  with  a an assymetric penalty of lasso type.
}
\examples{
data(ability.cov)
S <- ability.cov$cov
R <- stats::cov2cor(S)
d <- nrow(R)
L <- matrix(0,d,d)
U <- matrix(0.2,d,d)
diag(U) <- 0
res <- golazo(R,L=L,U=U)
Khat <- res$K
print(Khat)
}
\keyword{concentration}
\keyword{coordinate}
\keyword{descent,}
\keyword{matrix.}
