% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive.golazo.R
\name{positive.golazo}
\alias{positive.golazo}
\title{This is a wrapper for the positive GOLAZO problem.}
\usage{
positive.golazo(S, rho, tol = 1e-07, diagonal.pen = FALSE, verbose = TRUE)
}
\arguments{
\item{S}{Positive semidefinite matrix. This will be typically the sample covariance matrix but it can be somethink different in the dual likelihood computation or when the data follow the non-paranormal distribution.}

\item{rho}{the penalty on the positive entries of K (can be Inf).}

\item{tol}{The convergence tolerance (default tol=1e-7). The algorithm termininnates when teh dual gap (guaranteed to be nonnegative) is less than tol.}

\item{diagonal.pen}{if FALSE (default) the diagonal of K is not penalized.}

\item{verbose}{if TRUE (default) the output will be printed.}
}
\value{
K the optimal value of the concentration matrix

Sig the optimal value of the covariance matrix

it the number of iterations
}
\description{
The function simply running golazo() with L_ij=0 and U_ij=rho.
}
\examples{
data(ability.cov)
S <- ability.cov$cov
R <- stats::cov2cor(S)
d <- nrow(R)
res <- positive.golazo(R,rho=0.1)
Khat <- res$K
print(Khat)
}
\keyword{concentration}
\keyword{coordinate}
\keyword{descent,}
\keyword{matrix.}
