% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zmatrix.R
\name{Zmatrix}
\alias{Zmatrix}
\title{Computes the Z-matrix}
\usage{
Zmatrix(S)
}
\arguments{
\item{S}{a covariance matrix}
}
\description{
This function outputs the Z matrix, that is, the unique ultrametric matrix dominating S. This matrix is used to connstruct a starting point in teh GOLAZO algorithm when L=0 but U has strictly positive (off-diagonal entries).
}
\examples{
Zmatrix(diag(10))
}
\keyword{Z-matrix,}
\keyword{point}
\keyword{starting}
