% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golazo.R
\name{golazo}
\alias{golazo}
\title{Performs GOLAZO algorithm by optimizing the dual problem.}
\usage{
golazo(S, L, U, tol = 1e-07, verbose = TRUE)
}
\arguments{
\item{S}{Positive semidefinite matrix. This will be typically the sample covariance matrix but it can be somethink different in the dual likelihood computation or when the data follow the non-paranormal distribution.}

\item{L}{Matrix of lower penalties. It should have all entries non-positive (-Inf is a valid entry). The entries of L say how much negative entries of the inverse of Sigma are penalized. For GLASSO all entries should of L should be -rho and all entries of U should be rho (in both cases with zero diagonal). For positive GOLAZO L should be zero and U should be like for GLASSO.}

\item{U}{Matrix of upper penalties. It should have all entries non-negative (Inf is a valid entry). The entries of U say how much positive entries of the inverse of Sigma are penalized.}

\item{tol}{The convergence tolerance (default tol=1e-7). The algorithm termininnates when teh dual gap (guaranteed to be nonnegative) is less than tol.}

\item{verbose}{if TRUE (default) the output will be printed.}
}
\value{
K the optimal value of the concentration matrix

Sig the optimal value of the covariance matrix

it the number of iterations
}
\description{
This function implements a simple block-coordinate descent algorithm to find the maximum of the regularized
Gaussiann log-likelihood  with  a an assymetric penalty of lasso type.
}
\examples{
data(ability.cov)
S <- ability.cov$cov
R <- stats::cov2cor(S)
d <- nrow(R)
L <- matrix(0,d,d)
U <- matrix(0.2,d,d)
diag(U) <- 0
res <- golazo(R,L=L,U=U)
Khat <- res$K
print(Khat)
}
\keyword{concentration}
\keyword{coordinate}
\keyword{descent,}
\keyword{matrix.}
