% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive.golazo.R
\name{positive.golazo}
\alias{positive.golazo}
\title{This is a wrapper for the positive GOLAZO problem.}
\usage{
positive.golazo(S, rho, tol = 1e-07, diagonal.pen = FALSE, verbose = TRUE)
}
\arguments{
\item{S}{the sample covariance matrix}

\item{rho}{the penalty on the positive entries of K (can be Inf)}

\item{tol}{the convergence tolerance (default tol=1e-7)}

\item{diagonal.pen}{if FALSE (default) the diagonal of K is not penalized.}

\item{verbose}{if TRUE (default) the output will be printed.}
}
\value{
the optimal value of the concentration matrix

the number of iterations the algorithm needed to converge

the corresponding value of the log-likelihood
}
\description{
This is simply running golazo() with L_ij=0 and U_ij=rho.
}
\examples{
data(ability.cov)
S <- ability.cov$cov
R <- stats::cov2cor(S)
d <- nrow(R)
res <- positive.golazo(R,rho=0.1)
Khat <- res$K
print(Khat)
}
\keyword{concentration}
\keyword{coordinate}
\keyword{descent,}
\keyword{matrix.}
